Death
This code.. Was stroke inducing..

I kid you not.. 5 days of pure pain.. Trying to go through and modify classes that did not want to function at all... This was stroke inducing I now have depression..

On a semi serious note though. /getitems as a server admin to view a list of all semi-functional items..

Good luck, enjoy, I'm going to cry myself to sleep because of the absolute mess that the original program was.